Rename base class file name and fix imports
